+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+- Lang condition for `eachLabel`
+
 ## [0.4.2] - 2021-04-29
 
 ### Fixed